+++ include/nigiri/loader/init_finish.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <stdint.h>
+
 namespace nigiri {
 struct timetable;
 }
@@ -7,6 +9,6 @@ struct timetable;
 namespace nigiri::loader {
 
 void register_special_stations(timetable&);
-void finalize(timetable&, int const& no_profiles = 1);
+void finalize(timetable&, uint8_t const& no_profiles = 1);
 
 }  // namespace nigiri::loader
