+++ src/loader/build_footpaths.cc
@@ -417,7 +417,7 @@ void add_links_to_and_between_children(timetable& tt) {
   }
 }
 
-void write_footpaths(timetable& tt, int const& no_profiles = 1) {
+void write_footpaths(timetable& tt, uint8_t const& no_profiles = 1) {
   assert(tt.locations_.footpaths_out_.empty());
   assert(tt.locations_.footpaths_in_.empty());
   assert(tt.locations_.preprocessing_footpaths_out_.size() == tt.n_locations());
@@ -445,7 +445,7 @@ void write_footpaths(timetable& tt, int const& no_profiles = 1) {
   tt.locations_.preprocessing_footpaths_out_.clear();
 }
 
-void build_footpaths(timetable& tt, int const& no_profiles) {
+void build_footpaths(timetable& tt, uint8_t const& no_profiles) {
   add_links_to_and_between_children(tt);
   link_nearby_stations(tt);
   transitivize_footpaths(tt);
